+++ src/test/java/domain/pieces/ElephantTest.java
@@ -1,28 +1,25 @@
-package domain.piece;
+package domain.pieces;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.PieceOnRoute;
 import domain.Point;
 import domain.Team;
-import domain.pieces.EmptyPiece;
-import domain.pieces.Horse;
-import domain.pieces.Piece;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class HorseTest {
+class ElephantTest {
     @Test
     @DisplayName("피스가 이동할 수 있는 지점들을 전부 반환한다")
     void test_isAbleToArrive() {
         // given
-        Horse horse = new Horse(Team.CHO);
+        Elephant elephant = new Elephant(Team.CHO);
         Point startPoint = new Point(0, 0);
-        Point arrivalPoint = new Point(2, 1);
+        Point arrivalPoint = new Point(3, 2);
 
         // when
-        boolean actual = horse.isAbleToArrive(startPoint, arrivalPoint);
+        boolean actual = elephant.isAbleToArrive(startPoint, arrivalPoint);
 
         // then
         assertThat(actual).isTrue();
@@ -32,17 +29,18 @@ class HorseTest {
     @DisplayName("경로에 있는 모든 지점들을 반환한다")
     void test_getRoutePoints() {
         // given
-        Horse horse = new Horse(Team.CHO);
+        Elephant elephant = new Elephant(Team.CHO);
         Point startPoint = new Point(0, 0);
-        Point arrivalPoint = new Point(2, 1);
+        Point arrivalPoint = new Point(3, 2);
 
         // when
-        List<Point> routePoints = horse.getRoutePoints(startPoint, arrivalPoint);
+        List<Point> routePoints = elephant.getRoutePoints(startPoint, arrivalPoint);
 
         // then
         assertThat(routePoints).containsExactlyInAnyOrder(
                 new Point(1, 0),
-                new Point(2, 1)
+                new Point(2, 1),
+                new Point(3, 2)
         );
     }
 
@@ -50,48 +48,48 @@ class HorseTest {
     @DisplayName("경로 상 기물이 있으면 이동할 수 없다.")
     void test_isMovableWhenPieceOnRoute() {
         //given
-        Horse horse = new Horse(Team.CHO);
+        Elephant elephant = new Elephant(Team.CHO);
         Piece empty = new EmptyPiece();
-        PieceOnRoute pieceOnRoute = new PieceOnRoute(List.of(horse, empty));
+        PieceOnRoute pieceOnRoute = new PieceOnRoute(List.of(elephant, empty, empty));
 
         //when&then
-        assertThat(horse.isMovable(pieceOnRoute)).isFalse();
+        assertThat(elephant.isMovable(pieceOnRoute)).isFalse();
     }
 
     @Test
     @DisplayName("경로 상 기물이 없으면 이동할 수 없다.")
     void test_isMovable() {
         //given
-        Horse horse = new Horse(Team.CHO);
+        Elephant elephant = new Elephant(Team.CHO);
         Piece empty = new EmptyPiece();
-        PieceOnRoute pieceOnRoute = new PieceOnRoute(List.of(empty, empty));
+        PieceOnRoute pieceOnRoute = new PieceOnRoute(List.of(empty, empty, empty));
 
         //when&then
-        assertThat(horse.isMovable(pieceOnRoute)).isTrue();
+        assertThat(elephant.isMovable(pieceOnRoute)).isTrue();
     }
 
     @Test
     @DisplayName("도착점에 아군 기물이 있으면 이동할 수 없다.")
     void test_isMovableWhenPieceIsInMyTeam() {
         //given
-        Horse horse = new Horse(Team.CHO);
+        Elephant elephant = new Elephant(Team.CHO);
         Piece empty = new EmptyPiece();
-        PieceOnRoute pieceOnRoute = new PieceOnRoute(List.of(empty, horse));
+        PieceOnRoute pieceOnRoute = new PieceOnRoute(List.of(empty, empty, elephant));
 
         //when&then
-        assertThat(horse.isMovable(pieceOnRoute)).isFalse();
+        assertThat(elephant.isMovable(pieceOnRoute)).isFalse();
     }
 
     @Test
     @DisplayName("도착점에 아군 기물이 없으면 이동할 수 있다.")
     void test_isMovableWhenPieceIsInOtherTeam() {
         //given
-        Horse horseHan = new Horse(Team.HAN);
-        Horse horseCho = new Horse(Team.CHO);
+        Elephant elephantHan = new Elephant(Team.HAN);
+        Elephant elephantCho = new Elephant(Team.CHO);
         Piece empty = new EmptyPiece();
-        PieceOnRoute pieceOnRoute = new PieceOnRoute(List.of(empty, horseCho));
+        PieceOnRoute pieceOnRoute = new PieceOnRoute(List.of(empty, empty, elephantCho));
 
         //when&then
-        assertThat(horseHan.isMovable(pieceOnRoute)).isTrue();
+        assertThat(elephantHan.isMovable(pieceOnRoute)).isTrue();
     }
 }
