+++ src/test/java/domain/pieces/EmptyPieceTest.java
@@ -1,8 +1,6 @@
-package domain.piece;
+package domain.pieces;
 
 import domain.Team;
-import domain.pieces.EmptyPiece;
-import domain.pieces.Piece;
 import execptions.JanggiArgumentException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
